rename and differentiate between Authentication and Authorization controllers
