+++ controllers/AuthenticationController/checkEmail.js
