+++ controllers/AuthenticationController/resendVerification.js
