+++ controllers/AuthenticationController/verifyUser.js
