+++ controllers/AuthorizationController/index.js
@@ -1,0 +1,7 @@
+const authorize = require('./authorize');
+
+const AuthorizationController = {
+	authorize
+};
+
+module.exports = AuthorizationController
