+++ controllers/authorize.js
@@ -39,95 +39,7 @@ const User = require("../models/Users");
 const Permission = require("../models/Permissions");
 const Role = require("../models/Roles");
 
-/**
- * @desc Check if user is authorized to perform permission
- */
 module.exports.authorize = async (req, res) => {
-  let id;
-  let response = {
-    status: 401,
-    authorized: false,
-    message: "user is not authorized for this permission",
-  };
-
-  const { token, permission } = req.body;
-
-  if (!token) {
-    return res.status(401).json(response);
-  }
-
-  if (permission && !all_permissions.includes(permission)) {
-    return res.status(400).json({ status: 400, message: "invalid permission" });
-  }
-
-  try {
-    const decode = jwt.verify(token, process.env.JWT_SECRET);
-    id = decode.id;
-  } catch (error) {
-    return res.status(401).json(response);
-  }
-
-  const user = await User.findByPk(id, {
-    include: [
-      {
-        model: Permission,
-        as: "permissions",
-        attributes: ["name"],
-      },
-      {
-        model: Role,
-        as: "role",
-        attributes: ["name"],
-        include: [{ model: Permission, attributes: ["name"] }],
-      },
-    ],
-  });
-
-  if (user && !user.is_verified) {
-    return res.status(401).json({
-      status: 401,
-      authorized: false,
-      message: "user is not verified",
-    });
-  }
-
-  const userPermissions = user.permissions.map((permission) => permission.name);
-  const rolePermissions = user.role.permissions.map(
-    (permission) => permission.name,
-  );
-
-  const permissions = [...new Set([...userPermissions, ...rolePermissions])];
-
-  if (user && !permission) {
-    response = {
-      status: 200,
-      authorized: true,
-      message: "user is authenticated",
-      user: {
-        id,
-        role: user.role.name,
-		permissions,
-      },
-    };
-    return res.status(200).json(response);
-  }
-
-
-  if (permission && permissions.includes(permission)) {
-    response = {
-      status: 200,
-      authorized: true,
-      message: "user is authorized for this permission",
-      user: {
-        id,
-        permissions,
-        role: user.role.name,
-      },
-    };
-    return res.status(200).json(response);
-  }
-
-  return res.status(401).json(response);
 };
 
 module.exports.getAuthPermissions = async (req, res) => {
